refactor: 도메인 모델의 테스트 코드 패키지 분리
