+++ src/test/kotlin/blackjack/domain/card/CardDeckTest.kt
@@ -1,9 +1,5 @@
-package blackjack.domain
+package blackjack.domain.card
 
-import blackjack.domain.card.Card
-import blackjack.domain.card.CardDeck
-import blackjack.domain.card.CardNumber
-import blackjack.domain.card.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
