+++ src/test/kotlin/blackjack/domain/card/CardTest.kt
@@ -1,8 +1,5 @@
-package blackjack.domain
+package blackjack.domain.card
 
-import blackjack.domain.card.Card
-import blackjack.domain.card.CardNumber
-import blackjack.domain.card.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
