+++ src/test/kotlin/blackjack/domain/money/BetMoneyTest.kt
@@ -1,6 +1,5 @@
-package blackjack.domain
+package blackjack.domain.money
 
-import blackjack.domain.money.BetMoney
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
