+++ src/test/kotlin/blackjack/domain/participant/PlayerTest.kt
@@ -1,9 +1,8 @@
-package blackjack.domain
+package blackjack.domain.participant
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.CardNumber
 import blackjack.domain.card.Suit
-import blackjack.domain.participant.Player
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
